Bug fixes
1. Fixed a bug in which the program was entering an infinite loop when
the destination path was a descendant of the parent path.

2. Fixed a bug in which the the program was not able to correctly detect
the descendence of the destination directory if the program was executed
after changing the current directory to the executable's directory in
cmd and giving the destination as the current directory and source as
some ancestor of the current directory.
Refer
http://stackoverflow.com/questions/30932049/different-output-when-executing-from-cmd-and-codeblocks
for more info on this bug.

3. Also, converted the paths to their canonical versions before passing
them to the main function to enable symbolic links in the input paths.